Revert memo removal for selection ids
