+++ packages/block-editor/src/store/selectors.js
@@ -30,7 +30,6 @@ import {
 	checkAllowListRecursive,
 	checkAllowList,
 	getAllPatternsDependants,
-	getSelectedBlockClientIdsUnmemoized,
 } from './utils';
 import { orderBy } from '../utils/sorting';
 import { STORE_NAME } from './constants';
@@ -758,7 +757,39 @@ export function getSelectedBlocksInitialCaretPosition( state ) {
  * @return {Array} Multi-selected block client IDs.
  */
 export const getSelectedBlockClientIds = createSelector(
-	getSelectedBlockClientIdsUnmemoized,
+	( state ) => {
+		const { selectionStart, selectionEnd } = state.selection;
+
+		if ( ! selectionStart.clientId || ! selectionEnd.clientId ) {
+			return EMPTY_ARRAY;
+		}
+
+		if ( selectionStart.clientId === selectionEnd.clientId ) {
+			return [ selectionStart.clientId ];
+		}
+
+		// Retrieve root client ID to aid in retrieving relevant nested block
+		// order, being careful to allow the falsey empty string top-level root
+		// by explicitly testing against null.
+		const rootClientId = getBlockRootClientId(
+			state,
+			selectionStart.clientId
+		);
+
+		if ( rootClientId === null ) {
+			return EMPTY_ARRAY;
+		}
+
+		const blockOrder = getBlockOrder( state, rootClientId );
+		const startIndex = blockOrder.indexOf( selectionStart.clientId );
+		const endIndex = blockOrder.indexOf( selectionEnd.clientId );
+
+		if ( startIndex > endIndex ) {
+			return blockOrder.slice( endIndex, startIndex + 1 );
+		}
+
+		return blockOrder.slice( startIndex, endIndex + 1 );
+	},
 	( state ) => [
 		state.blocks.order,
 		state.selection.selectionStart.clientId,
@@ -821,10 +852,7 @@ export const getMultiSelectedBlocks = createSelector(
  * @return {?string} First block client ID in the multi-selection set.
  */
 export function getFirstMultiSelectedBlockClientId( state ) {
-	if ( ! hasMultiSelection( state ) ) {
-		return null;
-	}
-	return getSelectedBlockClientIdsUnmemoized( state )[ 0 ];
+	return getMultiSelectedBlockClientIds( state )[ 0 ] || null;
 }
 
 /**
@@ -836,11 +864,8 @@ export function getFirstMultiSelectedBlockClientId( state ) {
  * @return {?string} Last block client ID in the multi-selection set.
  */
 export function getLastMultiSelectedBlockClientId( state ) {
-	if ( ! hasMultiSelection( state ) ) {
-		return null;
-	}
-	const selectedClientIds = getSelectedBlockClientIdsUnmemoized( state );
-	return selectedClientIds[ selectedClientIds.length - 1 ];
+	const selectedClientIds = getMultiSelectedBlockClientIds( state );
+	return selectedClientIds[ selectedClientIds.length - 1 ] || null;
 }
 
 /**
@@ -867,12 +892,7 @@ export function isFirstMultiSelectedBlock( state, clientId ) {
  * @return {boolean} Whether block is in multi-selection set.
  */
 export function isBlockMultiSelected( state, clientId ) {
-	if ( isBlockSelected( state, clientId ) ) {
-		return false;
-	}
-	return (
-		getSelectedBlockClientIdsUnmemoized( state ).indexOf( clientId ) !== -1
-	);
+	return getMultiSelectedBlockClientIds( state ).indexOf( clientId ) !== -1;
 }
 
 /**
@@ -984,7 +1004,7 @@ export function __unstableIsSelectionCollapsed( state ) {
 }
 
 export function __unstableSelectionHasUnmergeableBlock( state ) {
-	return getSelectedBlockClientIdsUnmemoized( state ).some( ( clientId ) => {
+	return getSelectedBlockClientIds( state ).some( ( clientId ) => {
 		const blockName = getBlockName( state, clientId );
 		const blockType = getBlockType( blockName );
 		return ! blockType.merge;
@@ -2918,7 +2938,7 @@ export const isGroupable = createRegistrySelector(
 			const groupingBlockName = getGroupingBlockName();
 			const _clientIds = clientIds?.length
 				? clientIds
-				: getSelectedBlockClientIdsUnmemoized( state );
+				: getSelectedBlockClientIds( state );
 			const rootClientId = _clientIds?.length
 				? getBlockRootClientId( state, _clientIds[ 0 ] )
 				: undefined;
